Switch Case Optimisation
- Changed the way the new colors for the percentiles are determined.
- Function now returns long
- WHITE and OFF can now also be returned on a scale between 0 and 100

+++ src/Excelsior_Light.h
@@ -19,9 +19,9 @@ class ELight{
     public:
         ELight(int red, int green, int blue, int signal);
         void changeLightDelay(int delay);
-        int  readSensor() {readSensor(WHITE,false)};
-        int  readSensor(int color) {readSensor(color,false)};
-        int  readSensor(int color, bool percent);
+        long readSensor() {readSensor(WHITE,false)};
+        long readSensor(int color) {readSensor(color,false)};
+        long readSensor(int color, bool percent);
     
     private:        
         int _ports[4];
